Added Service Worker
Finally got the service worker registered. Was having a hard time with it, but seems to be working now after trial and error. I believe the issue was I was trying to call the imported useServiceWorker() function directly in layout.tsx. Now I am calling it as a child component of layout.tsx (it returns null as there is nothing to render visibly)